waranty/requirements.txt
asgiref==3.8.1
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Django==5.1.5
django-rest-knox==5.0.2
djangorestframework==3.15.2
dynaconf==3.2.6
hvac==2.3.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
knox==0.1.14
loguru==0.7.3
MarkupSafe==3.0.2
mysqlclient==2.2.7
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.1
setuptools==75.8.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
validators==0.34.0
